Added more intervals
