+++ src/Limit.php
@@ -6,7 +6,6 @@ namespace Saloon\RateLimiter;
 
 use Closure;
 use ReflectionClass;
-use DateTimeImmutable;
 use Saloon\Helpers\Date;
 use InvalidArgumentException;
 use Saloon\Contracts\Request;
@@ -14,9 +13,12 @@ use Saloon\Contracts\Response;
 use Saloon\Contracts\Connector;
 use Saloon\RateLimiter\Exceptions\LimitException;
 use Saloon\RateLimiter\Contracts\RateLimiterStore;
+use Saloon\RateLimiter\Traits\HasIntervals;
 
 class Limit
 {
+    use HasIntervals;
+
     /**
      * Connector string
      *
@@ -62,21 +64,6 @@ class Limit
      * @var int|null
      */
     protected ?int $expiryTimestamp = null;
-
-    /**
-     * The number of seconds it will take to release the rate limit after it has
-     * been reached.
-     *
-     * @var int
-     */
-    protected int $releaseInSeconds;
-
-    /**
-     * Optional time to live key to specify the time in the default key.
-     *
-     * @var string|null
-     */
-    protected ?string $timeToLiveKey = null;
 
     /**
      * Determines if a limit has been manually exceeded.
@@ -146,6 +133,12 @@ class Limit
         return $this->hits >= ($threshold * $this->allow);
     }
 
+    /**
+     * Hit the limit
+     *
+     * @param int $amount
+     * @return $this
+     */
     public function hit(int $amount = 1): static
     {
         if (! $this->hasExceeded()) {
@@ -155,6 +148,12 @@ class Limit
         return $this;
     }
 
+    /**
+     * Set the limit as exceeded
+     *
+     * @param $releaseInSeconds
+     * @return void
+     */
     public function exceeded($releaseInSeconds = null): void
     {
         $this->exceeded = true;
@@ -166,6 +165,11 @@ class Limit
         }
     }
 
+    /**
+     * Get the hits
+     *
+     * @return int
+     */
     public function getHits(): int
     {
         return $this->hits;
@@ -209,6 +213,8 @@ class Limit
     }
 
     /**
+     * Get the expiry timestamp
+     *
      * @return int|null
      */
     public function getExpiryTimestamp(): ?int
@@ -217,6 +223,8 @@ class Limit
     }
 
     /**
+     * Set the expiry timestamp
+     *
      * @param int|null $expiryTimestamp
      * @return Limit
      */
@@ -238,59 +246,6 @@ class Limit
         return $this->setExpiryTimestamp($this->getCurrentTimestamp() + $seconds);
     }
 
-    public function everySeconds(int $seconds, ?string $timeToLiveKey = null): static
-    {
-        $this->releaseInSeconds = $seconds;
-        $this->timeToLiveKey = $timeToLiveKey;
-
-        return $this;
-    }
-
-    public function everyMinute(): static
-    {
-        return $this->everySeconds(60);
-    }
-
-    public function everyFiveMinutes(): static
-    {
-        return $this->everySeconds(60 * 5);
-    }
-
-    public function everyThirtyMinutes(): static
-    {
-        return $this->everySeconds(60 * 30);
-    }
-
-    public function everyHour(): static
-    {
-        return $this->everySeconds(60 * 60);
-    }
-
-    public function everySixHours(): static
-    {
-        return $this->everySeconds(60 * 60 * 6);
-    }
-
-    public function everyTwelveHours(): static
-    {
-        return $this->everySeconds(60 * 60 * 12);
-    }
-
-    public function everyDay(): static
-    {
-        return $this->everySeconds(60 * 60 * 24);
-    }
-
-    public function untilMidnightTonight(): static
-    {
-        $tomorrowTimestamp = (new DateTimeImmutable('tomorrow'))->getTimestamp();
-
-        return $this->everySeconds(
-            seconds: $tomorrowTimestamp - $this->getCurrentTimestamp(),
-            timeToLiveKey: 'midnight'
-        );
-    }
-
     /**
      * Get the remaining time in seconds
      *
@@ -301,26 +256,52 @@ class Limit
         return (int)round($this->getExpiryTimestamp() - $this->getCurrentTimestamp());
     }
 
+    /**
+     * Get the release time in seconds
+     *
+     * @return int
+     */
     public function getReleaseInSeconds(): int
     {
         return $this->releaseInSeconds;
     }
 
+    /**
+     * Check if the limit has been exceeded
+     *
+     * @return bool
+     */
     public function hasExceeded(): bool
     {
         return $this->exceeded;
     }
 
-    public function validate()
+    /**
+     * Validate the limit
+     *
+     * @return void
+     */
+    public function validate(): void
     {
         // Todo: Validate we have allow and releaseInSeconds
     }
 
+    /**
+     * Check if the limit uses a response
+     *
+     * @return bool
+     */
     public function usesResponse(): bool
     {
         return isset($this->responseHandler);
     }
 
+    /**
+     * Handle a response on the limit
+     *
+     * @param \Saloon\Contracts\Response $response
+     * @return void
+     */
     public function handleResponse(Response $response): void
     {
         if (! $this->usesResponse()) {
@@ -418,15 +399,5 @@ class Limit
         }
 
         return $this;
-    }
-
-    /**
-     * Get the current timestamp
-     *
-     * @return int
-     */
-    protected function getCurrentTimestamp(): int
-    {
-        return Date::now()->toDateTime()->getTimestamp();
     }
 }
